+++ lib/core/src/display/shape/text/msdf_frag.glsl
