+++ lib/core/src/display/shape/text/msdf_vert.glsl
